use google oauth2 for social authentication

+++ controllers/api/v1/posts_api.js
@@ -30,7 +30,7 @@ module.exports.destroy = async function(request,response){
 
         let post=  await Post.findById(request.params.id);
              //.id means converting the object id into string  
-        // if(post.user == request.user.id){
+        if(post.user == request.user.id){
             post.remove();
 
               await Comment.deleteMany({ post:request.params.id});
@@ -52,7 +52,11 @@ module.exports.destroy = async function(request,response){
               });   
        
            
- 
+            }else{
+              return response.json(401,{
+                message:"You cannot delete this post"
+              })
+            }
 
 
     }catch(error){
